Make ViewControllerNavigationAction a private class
